+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Search pane
 - `search_button` config option for search pane
 - Configurable `Save()` keybind in the navigation section
 - Added `SeekToStart` global action to seek to the beginning of the currently playing track
+- `DeleteFromPlaylist` action to delete songs in browsers panes from selected playlist
 
 ### Changed